requirements.txt
pywavefront
PyGLM
Pillow
numpy
glfw
PyOpenGL
